refactor(crew): split Pure Static Methods region into CrewReservationManager.SeatMatch.cs
